requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
future==0.18.3
imageio==2.25.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
numpy-indexed==0.3.7
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
